Implement queueing at the leader
This PR implements a queue at the leader that will wait for transactions being
sent by clients. A first transaction in the queue triggers the leader into waiting
for more transaction during a given timeout. When the timeout expires, the
leader will collect all transactions and propose them to the other nodes
for inclusion in omniledger's state.

Also started documentation for the parts of omniledger that are implemented.