Implement online verification for darcs
This PR allows verification of darcs when the signature path is empty,
but there exists a digest of the signature path. The server, given an
appropriate callbacks to find darcs, is able to find the missing path
and use that for the verification.  This is only the first step, more
improvements are planned in the future, which we describe in #33.