Add the "sign" action to darc
When darc A delegates an action to another darc B. We will evaluate the
sign expression on darc B to check whether the signatures are correct.
The "sign" expression can, of course, be further delegations, then the
evaluator will recursively check the chain of delegation until it find
an actual signing key.