Generalise Request to support evolution
We change the message that signers must sign to evolve a darc to be a
digest of a Request struct, where before it was a digest of the darc.
This approach generalises our Request struct to works for every kind of
request. It also simplifies verification because we can re-use many of
the existing functions from Darc.Verify.